+++ CHANGELOG.md
@@ -18,6 +18,9 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * The `flood_map.iterative` method now generates an initial guess using the `nmad` method and then runs with a maximum step size of 3 instead of the default 0.5.
 * the known water threshold used to determine perennial water when creating flood maps will be calculated by default, or if `NaN`, using `flood_map.get_pw_threshold`
 
+### Fixed
+* Reverted the special handling of nan values introduced in v0.5.2, now that GDAL v3.7.0 has been released.
+
 ## [0.5.2]
 
 ### Added